Refactored soz_vers/.
